+++ R/params.R
@@ -5,6 +5,11 @@
 #' @param ... Unused.
 #' @param flesh An atomic vector
 #' @param skeleton An nlist object.
+#' @param scalar A logical scalar specifying whether to include 
+#' all parameters (NA), only scalars (TRUE) or all parameters
+#' except scalars (FALSE).
+#' @param terms A flag specifying whether to return the parameter name
+#' for each term element.
 #'
 #' @name params
 NULL